Fixed so meta infos are removed with pages

+++ src/java/org/infoglue/cms/controllers/kernel/impl/simple/ContentController.java
@@ -259,7 +259,7 @@ public class ContentController extends BaseController
         beginTransaction(db);
 		try
         {		
-	    	delete(contentVO, db);
+	    	delete(contentVO, db, false, false);
 	    	
 	    	commitTransaction(db);
         }
@@ -278,12 +278,21 @@ public class ContentController extends BaseController
 
     }  
     
-    
+
 	/**
 	 * This method deletes a content and also erases all the children and all versions.
 	 */
 	    
 	public void delete(ContentVO contentVO, Database db) throws ConstraintException, SystemException, Exception
+	{
+	    delete(contentVO, db, false, false);
+	}
+	
+	/**
+	 * This method deletes a content and also erases all the children and all versions.
+	 */
+	    
+	public void delete(ContentVO contentVO, Database db, boolean skipRelationCheck, boolean skipServiceBindings) throws ConstraintException, SystemException, Exception
 	{
 		Content content = getContentWithId(contentVO.getContentId(), db);
 	    Content parent = content.getParentContent();
@@ -295,13 +304,13 @@ public class ContentController extends BaseController
 			    Content candidate = (Content)childContentIterator.next();
 			    if(candidate.getId().equals(contentVO.getContentId()))
 			    {
-			        deleteRecursive(content, childContentIterator, db);
+			        deleteRecursive(content, childContentIterator, db, skipRelationCheck, skipServiceBindings);
 			    }
 			}
 		}
 		else
 		{
-		    deleteRecursive(content, null, db);
+		    deleteRecursive(content, null, db, skipRelationCheck, skipServiceBindings);
 		}
 	}        
 
@@ -309,18 +318,21 @@ public class ContentController extends BaseController
 	 * Recursively deletes all contents and their versions. Also updates related entities about the change.
 	 */
 	
-    private static void deleteRecursive(Content content, Iterator parentIterator, Database db) throws ConstraintException, SystemException, Exception
+    private static void deleteRecursive(Content content, Iterator parentIterator, Database db, boolean skipRelationCheck, boolean skipServiceBindings) throws ConstraintException, SystemException, Exception
     {
-        List referenceBeanList = RegistryController.getController().getReferencingObjectsForContent(content.getId(), db);
-		if(referenceBeanList != null && referenceBeanList.size() > 0)
-			throw new ConstraintException("ContentVersion.stateId", "3305");
-		
+        if(!skipRelationCheck)
+        {
+	        List referenceBeanList = RegistryController.getController().getReferencingObjectsForContent(content.getId(), db);
+			if(referenceBeanList != null && referenceBeanList.size() > 0)
+				throw new ConstraintException("ContentVersion.stateId", "3305");
+        }
+        
         Collection children = content.getChildren();
 		Iterator childrenIterator = children.iterator();
 		while(childrenIterator.hasNext())
 		{
 			Content childContent = (Content)childrenIterator.next();
-			deleteRecursive(childContent, childrenIterator, db);   			
+			deleteRecursive(childContent, childrenIterator, db, skipRelationCheck, skipServiceBindings);   			
    		}
 		content.setChildren(new ArrayList());
    		
@@ -328,7 +340,8 @@ public class ContentController extends BaseController
 	    {		 
 			ContentVersionController.getContentVersionController().deleteVersionsForContent(content, db);    	
 			
-			ServiceBindingController.deleteServiceBindingsReferencingContent(content, db);
+			if(!skipServiceBindings)
+			    ServiceBindingController.deleteServiceBindingsReferencingContent(content, db);
 			
 			if(parentIterator != null) 
 			    parentIterator.remove();
@@ -1184,7 +1197,7 @@ public class ContentController extends BaseController
         return languages;
     }        
 
-
+    
 	/**
 	 * This method returns the bound contents based on a servicebinding.
 	 */
@@ -1199,48 +1212,8 @@ public class ContentController extends BaseController
 		
 		try
 		{
-			ServiceBinding serviceBinding = ServiceBindingController.getServiceBindingWithId(serviceBindingId, db);
-	        
-			if(serviceBinding != null)
-			{
-				ServiceDefinition serviceDefinition = serviceBinding.getServiceDefinition();
-				if(serviceDefinition != null)
-				{
-					String serviceClassName = serviceDefinition.getClassName();
-					BaseService service = (BaseService)Class.forName(serviceClassName).newInstance();
-	        		 
-					HashMap arguments = new HashMap();
-					arguments.put("method", "selectContentListOnIdList");
-	            		
-					List qualifyerList = new ArrayList();
-					Collection qualifyers = serviceBinding.getBindingQualifyers();
-
-					qualifyers = sortQualifyers(qualifyers);
-
-					Iterator iterator = qualifyers.iterator();
-					while(iterator.hasNext())
-					{
-						Qualifyer qualifyer = (Qualifyer)iterator.next();
-						HashMap argument = new HashMap();
-						argument.put(qualifyer.getName(), qualifyer.getValue());
-						qualifyerList.add(argument);
-					}
-					arguments.put("arguments", qualifyerList);
-	        		
-					List contents = service.selectMatchingEntities(arguments);
-	        		
-					if(contents != null)
-					{
-						Iterator i = contents.iterator();
-						while(i.hasNext())
-						{
-							ContentVO candidate = (ContentVO)i.next();
-							result.add(candidate);        		
-						}
-					}
-				}
-			}
-	       	  
+		    result = getBoundContents(db, serviceBindingId);
+		    
 			commitTransaction(db);
 		}
 		catch(Exception e)
@@ -1250,6 +1223,60 @@ public class ContentController extends BaseController
 			throw new SystemException(e.getMessage());
 		}
 		
+		return result;
+	}
+
+
+	/**
+	 * This method returns the bound contents based on a servicebinding.
+	 */
+	
+	public static List getBoundContents(Database db, Integer serviceBindingId) throws SystemException, Exception
+	{
+		List result = new ArrayList();
+		
+		ServiceBinding serviceBinding = ServiceBindingController.getServiceBindingWithId(serviceBindingId, db);
+        
+		if(serviceBinding != null)
+		{
+			ServiceDefinition serviceDefinition = serviceBinding.getServiceDefinition();
+			if(serviceDefinition != null)
+			{
+				String serviceClassName = serviceDefinition.getClassName();
+				BaseService service = (BaseService)Class.forName(serviceClassName).newInstance();
+        		 
+				HashMap arguments = new HashMap();
+				arguments.put("method", "selectContentListOnIdList");
+            		
+				List qualifyerList = new ArrayList();
+				Collection qualifyers = serviceBinding.getBindingQualifyers();
+
+				qualifyers = sortQualifyers(qualifyers);
+
+				Iterator iterator = qualifyers.iterator();
+				while(iterator.hasNext())
+				{
+					Qualifyer qualifyer = (Qualifyer)iterator.next();
+					HashMap argument = new HashMap();
+					argument.put(qualifyer.getName(), qualifyer.getValue());
+					qualifyerList.add(argument);
+				}
+				arguments.put("arguments", qualifyerList);
+        		
+				List contents = service.selectMatchingEntities(arguments);
+        		
+				if(contents != null)
+				{
+					Iterator i = contents.iterator();
+					while(i.hasNext())
+					{
+						ContentVO candidate = (ContentVO)i.next();
+						result.add(candidate);        		
+					}
+				}
+			}
+		}
+	       	  		
 		return result;
 	}
